% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generators.R
\name{generate_continuous}
\alias{generate_continuous}
\title{Title}
\usage{
generate_continuous(
  n_studies,
  study_sample_size,
  n_predictors = 12,
  sigmas,
  intercepts_data = NULL,
  min_study_id = 1,
  predictor_intercepts = "study"
)
}
\arguments{
\item{n_studies}{number of studies}

\item{study_sample_size}{sample size for all studies}

\item{n_predictors}{number of predictors}

\item{sigmas}{output from get_sigmas with error variances and other model terms}

\item{intercepts_data}{a vector contianing study intercepts. This can be used to generate data from existing studies}

\item{min_study_id}{the minimum study ID to use. When generating data for new studies this will need to exceed the ID used in train data.}

\item{predictor_intercepts}{Either 'study' or 'random'. If study, the intercepts passed to the function that generates the predictor data are study intercepts. If random the interceps passed are randomly generated (or existing preductor intercepts if intercept data is provided)
Use study when data is generated with a correlation betweens study interceps and predictos. Use random if data is generated with predictors correlated within study. If intercepts are iid accross studies this option will make no difference.}
}
\value{
returns a dataframe of simualated data with outcome y and predictors x1 ... x{n_predictors}
}
\description{
Title
}
\examples{

I should refactor so n_predictos, and predictor interceps are part of sigmas. That way all arguments that are unchanged between reps, and between train and test data are contained in sigmas.

}
