% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipdma_prediction_pipeline_dynamic.R
\name{ipdma_prediction_pipeline_dynamic}
\alias{ipdma_prediction_pipeline_dynamic}
\title{Pipeline for developing and validating a prediction model by study for an IPD Meta-analysis with training and test data.}
\usage{
ipdma_prediction_pipeline_dynamic(
  test_data,
  model_function,
  train_data,
  evaluate_performance
)
}
\arguments{
\item{test_data}{data used to validate the model}

\item{train_data}{data used to fit the model}

\item{evaluate_performance}{a function that evaluates the performance of the model for a single study}

\item{model}{a fitted model}
}
\description{
Pipeline for developing and validating a prediction model by study for an IPD Meta-analysis with training and test data.
}
