% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_investigations.R
\name{sim_rep_continuous_inv}
\alias{sim_rep_continuous_inv}
\title{Title}
\usage{
sim_rep_continuous_inv(
  model_function_list,
  n_studies,
  study_sample_size_train,
  study_sample_size_test,
  sigma_e,
  sigma_u,
  sim_rep
)
}
\arguments{
\item{sim_rep}{data from the simulation rep to be investigated}
}
\description{
Title
}
