% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipdma_prediction_pipeline.R
\name{ipdma_prediction_pipeline}
\alias{ipdma_prediction_pipeline}
\title{Pipeline for developing and validating a prediction model by study for an IPD Meta-analysis with training and test data.}
\usage{
ipdma_prediction_pipeline(test_data, model, evaluate_performance)
}
\arguments{
\item{test_data}{data used to validate the model}

\item{model}{a fitted model}

\item{evaluate_performance}{a function that evaluates the performance of the model for a single study}

\item{train_data}{data used to fit the model}
}
\description{
Pipeline for developing and validating a prediction model by study for an IPD Meta-analysis with training and test data.
}
